+++ src/main/java/com/techoble/reviewer/controller/ReviewerController.java
@@ -1,7 +1,7 @@
 package com.techoble.reviewer.controller;
 
-import com.techoble.reviewer.domain.CrewsDto;
-import com.techoble.reviewer.service.MatchingService;
+import com.techoble.reviewer.dto.CrewsDto;
+import com.techoble.reviewer.service.ReviewerService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -9,11 +9,12 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 @Controller
-public class MatchingController {
-    private final MatchingService matchingService;
+public class ReviewerController {
 
-    public MatchingController(final MatchingService matchingService) {
-        this.matchingService = matchingService;
+    private final ReviewerService reviewerService;
+
+    public ReviewerController(final ReviewerService reviewerService) {
+        this.reviewerService = reviewerService;
     }
 
     @GetMapping("/crews")
@@ -23,7 +24,7 @@ public class MatchingController {
 
     @GetMapping("/api/crews")
     public ResponseEntity<CrewsDto> findCrews() {
-        CrewsDto crews = matchingService.findCrews();
+        CrewsDto crews = reviewerService.findCrews();
         if (crews.getCrews().isEmpty()) {
             return ResponseEntity.noContent().build();
         }
@@ -32,7 +33,7 @@ public class MatchingController {
 
     @PostMapping("/api/crews")
     public ResponseEntity<Void> register(@RequestParam String name) {
-        matchingService.add(name);
+        reviewerService.add(name);
         return ResponseEntity.ok().build();
     }
 }
